# src/Umbraco.Community.SchemeWeaver.Mcp/src/umbraco-api/api/schemeweaver-openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "SchemeWeaver Management API",
    "description": "SchemeWeaver endpoints extracted from the Umbraco management API document (https://localhost:44308/umbraco/openapi/management.json)",
    "version": "Latest"
  },
  "paths": {
    "/umbraco/management/api/v1/schemeweaver/content-types": {
      "get": {
        "tags": [
          "SchemeWeaver"
        ],
        "operationId": "GetSchemeweaverContentTypes",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "The resource is protected and requires an authentication token"
          }
        },
        "security": [
          {
            "Backoffice-User": []
          }
        ]
      }
    },
    "/umbraco/management/api/v1/schemeweaver/content-types/{alias}/properties": {
      "get": {
        "tags": [
          "SchemeWeaver"
        ],
        "operationId": "GetSchemeweaverContentTypesByAliasProperties",
        "parameters": [
          {
            "name": "alias",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "The resource is protected and requires an authentication token"
          }
        },
        "security": [
          {
            "Backoffice-User": []
          }
        ]
      }
    },
    "/umbraco/management/api/v1/schemeweaver/content-types/{contentTypeAlias}/properties/{propertyAlias}/block-suggest": {
      "post": {
        "tags": [
          "SchemeWeaver"
        ],
        "operationId": "PostSchemeweaverContentTypesByContentTypeAliasPropertiesByPropertyAliasBlockSuggest",
        "parameters": [
          {
            "name": "contentTypeAlias",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "propertyAlias",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Umb-Notifications": {
                "description": "The list of notifications produced during the request.",
                "schema": {
                  "type": [
                    "null",
                    "array"
                  ],
                  "items": {
                    "$ref": "#/components/schemas/NotificationHeaderModel"
                  }
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BlockMappingSuggestion"
                  }
                }
              }
            }
          },
          "401": {
            "description": "The resource is protected and requires an authentication token"
          }
        },
        "security": [
          {
            "Backoffice-User": []
          }
        ]
      }
    },
    "/umbraco/management/api/v1/schemeweaver/content-types/{contentTypeAlias}/properties/{propertyAlias}/block-types": {
      "get": {
        "tags": [
          "SchemeWeaver"
        ],
        "operationId": "GetSchemeweaverContentTypesByContentTypeAliasPropertiesByPropertyAliasBlockTypes",
        "parameters": [
          {
            "name": "contentTypeAlias",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "propertyAlias",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BlockElementTypeInfo"
                  }
                }
              }
            }
          },
          "401": {
            "description": "The resource is protected and requires an authentication token"
          }
        },
        "security": [
          {
            "Backoffice-User": []
          }
        ]
      }
    },
    "/umbraco/management/api/v1/schemeweaver/generate-content-type": {
      "post": {
        "tags": [
          "SchemeWeaver"
        ],
        "operationId": "PostSchemeweaverGenerateContentType",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentTypeGenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Umb-Notifications": {
                "description": "The list of notifications produced during the request.",
                "schema": {
                  "type": [
                    "null",
                    "array"
                  ],
                  "items": {
                    "$ref": "#/components/schemas/NotificationHeaderModel"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "headers": {
              "Umb-Notifications": {
                "description": "The list of notifications produced during the request.",
                "schema": {
                  "type": [
                    "null",
                    "array"
                  ],
                  "items": {
                    "$ref": "#/components/schemas/NotificationHeaderModel"
                  }
                }
              }
            }
          },
          "401": {
            "description": "The resource is protected and requires an authentication token"
          }
        },
        "security": [
          {
            "Backoffice-User": []
          }
        ]
      }
    },
...
    "/umbraco/management/api/v1/schemeweaver/schema-types": {
      "get": {
        "tags": [
          "SchemeWeaver"
        ],
        "operationId": "GetSchemeweaverSchemaTypes",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SchemaTypeInfo"
                  }
                }
              }
            }
          },
          "401": {
            "description": "The resource is protected and requires an authentication token"
          }
        },
        "security": [
          {
            "Backoffice-User": []
          }
        ]
      }
    },
    "/umbraco/management/api/v1/schemeweaver/schema-types/{name}/properties": {
      "get": {
        "tags": [
          "SchemeWeaver"
        ],
        "operationId": "GetSchemeweaverSchemaTypesByNameProperties",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ranked",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RankedSchemaPropertyInfo"
                  }
                }
              }
            }
          },
          "401": {
            "description": "The resource is protected and requires an authentication token"
          }
        },
        "security": [
          {
            "Backoffice-User": []
          }
        ]
      }
    },
    "/umbraco/management/api/v1/schemeweaver/server-context": {
      "get": {
        "tags": [
          "SchemeWeaver"
        ],
        "operationId": "GetSchemeweaverServerContext",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerContextDto"
                }
              }
            }
          },
          "401": {
            "description": "The resource is protected and requires an authentication token"
          }
        },
        "security": [
          {
            "Backoffice-User": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "BlockElementPropertyInfo": {
        "required": [
          "alias",
          "name",
          "editorAlias",
          "nestedBlockElementTypes"
        ],
        "type": "object",
        "properties": {
          "alias": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "editorAlias": {
            "type": "string"
          },
          "nestedBlockElementTypes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlockElementTypeInfo"
            }
          }
        }
      },
      "BlockElementTypeInfo": {
        "required": [
          "alias",
          "name",
          "properties",
          "propertyInfos"
        ],
        "type": "object",
        "properties": {
          "alias": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "properties": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "propertyInfos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlockElementPropertyInfo"
            }
          }
        }
      },
      "BlockMappingSuggestion": {
        "required": [
          "schemaProperty",
          "confidence",
          "routes"
        ],
        "type": "object",
        "properties": {
          "schemaProperty": {
            "type": "string"
          },
          "confidence": {
            "type": "integer",
            "format": "int32"
          },
          "routes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlockRouteSuggestion"
            }
          }
        }
      },
      "BlockRoutePropertyMappingSuggestion": {
        "required": [
          "schemaProperty",
          "contentProperty"
        ],
        "type": "object",
        "properties": {
          "schemaProperty": {
            "type": "string"
          },
          "contentProperty": {
            "type": "string"
          },
          "wrapInType": {
            "type": [
              "null",
              "string"
            ]
          },
          "wrapInProperty": {
            "type": [
              "null",
              "string"
            ]
          },
          "routes": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/BlockRouteSuggestion"
            }
          }
        }
      },
      "BlockRouteSuggestion": {
        "required": [
          "blockAlias",
          "nestedSchemaType",
          "confidence",
          "propertyMappings"
        ],
        "type": "object",
        "properties": {
          "blockAlias": {
            "type": "string"
          },
          "nestedSchemaType": {
            "type": "string"
          },
          "confidence": {
            "type": "integer",
            "format": "int32"
          },
          "propertyMappings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlockRoutePropertyMappingSuggestion"
            }
          }
        }
      },
      "ContentTypeGenerationRequest": {
        "required": [
          "schemaTypeName",
          "documentTypeName",
          "documentTypeAlias",
          "selectedProperties",
          "propertyGroupName"
        ],
        "type": "object",
        "properties": {
          "schemaTypeName": {
            "type": "string"
          },
          "documentTypeName": {
            "type": "string"
          },
          "documentTypeAlias": {
            "type": "string"
          },
          "selectedProperties": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "propertyGroupName": {
            "type": "string"
          }
        }
      },
      "EventMessageTypeModel": {
        "enum": [
          "Default",
          "Info",
          "Error",
          "Success",
          "Warning"
        ]
      },
...
      "NotificationHeaderModel": {
        "required": [
          "message",
          "category",
          "type"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/EventMessageTypeModel"
          }
        }
      },
...
      "RankedSchemaPropertyInfo": {
        "required": [
          "confidence",
          "isPopular",
          "name",
          "propertyType",
          "isRequired",
          "acceptedTypes",
          "isComplexType"
        ],
        "type": "object",
        "properties": {
          "confidence": {
            "type": "integer",
            "format": "int32"
          },
          "isPopular": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "propertyType": {
            "type": "string"
          },
          "isRequired": {
            "type": "boolean"
          },
          "acceptedTypes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isComplexType": {
            "type": "boolean"
          }
        }
      },
...
      "SchemaTypeInfo": {
        "required": [
          "name",
          "propertyCount"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "parentTypeName": {
            "type": [
              "null",
              "string"
            ]
          },
          "propertyCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ServerContextDto": {
        "required": [
          "hasPublishedContent",
          "isTestHost"
        ],
        "type": "object",
        "properties": {
          "hasPublishedContent": {
            "type": "boolean"
          },
          "isTestHost": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Backoffice-User": {
        "type": "oauth2",
        "description": "Umbraco Authentication",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/umbraco/management/api/v1/security/back-office/authorize",
            "tokenUrl": "/umbraco/management/api/v1/security/back-office/token",
            "scopes": {}
          }
        }
      }
    }
  }
}